Use layer slicer in viewer model (#5377)
# Description
This uses the recently introduced `LayerSlicer` class to allow us to
perform slicing in some cases, and to easily enable async slicing in
those cases. There may still be bugs or performance issues with async
slicing, but merging this PR would at least allow us to test, measure,
and make progress. By default, slicing is synchronous (due to
`LayerSlicer._force_sync = True` in `LayerSlicer.__init__`) and behavior
should be identical to `main`. Performance should also be unaffected in
this case.

Asynchronous slicing can be enabled by executing
`viewer._layer_slicer._force_sync = False` before slicing occurs either
due to changes in `Dims`, adding layers, or panning/zooming 2D views of
multi-scale image layers. We need to make this friendlier to enable,
either by using the existing experimental async flag in
preferences/settings or by defining a new one. But I think that can
happen in another PR.

The big long-term design change here is that there's some separation
between slicing and layers. Previously, slicing could be achieved by
updating some private state on the layers and calling `Layer.refresh`.
Now the intent is to have all slicing go through the `ViewerModel`,
which delegates slicing to `LayerSlicer`, which then slices the layers
by combining state from `Dims` and each layer instance (potentially on a
separate thread). Each slice result is passed to observers of the
`LayerSlicer.ready` event. In napari, the only observer of that event is
`QtViewer`, which performs some post-slicing tasks that need to be
handled on the main thread (e.g. updating vispy layers).

This integration roughly follows the [description in
NAP-4](https://napari.org/dev/naps/4-async-slicing.html#detailed-description),
but with a few differences, some of which might be temporary.
- Instead of passing the slice response directly to vispy, we update the
slice state on the layer and notify vispy of the change. We need to
update the layer slice state for other reasons (e.g. `_get_value`), so
this keeps things simpler for now.
- Instead of a static `Layer.get_slice`, we made each slice request
callable.

Temporarily, this makes slicing logic in general more complex because
some layers support async slicing and some don't. But in the long term,
the separation of slicing from layers should provide a much needed
separation of concerns.

## Type of change
- [x] New feature (non-breaking change which adds functionality)

This makes one public API addition which is `Layer.events.reload`, which
is intended to be a long term replacement for `Layer.refresh` that
supports both sync and async slicing.

# References
Closes #5236 

# How has this been tested?
- [x] all existing tests pass with my change