Remove catch_warnings in slicing (#5603)
# Description
This removes use of `catch_warnings` in slicing because [using that
context manager has the side of effect of forgetting any previously
raised warnings](https://github.com/python/cpython/issues/73858). Using
those in slicing means that any warnings raised in slicing are shown
every time, whereas [the default behavior for Python is to only show the
first time per
location](https://docs.python.org/3/library/warnings.html#the-warnings-filter).

The removal of `catch_warnings` in `Image._update_thumbnail` is
justified because [napari requires scipy
>=1.4.1](https://github.com/andy-sweet/napari/blob/196fb164ba12402dd10869fa8e4fd62ed8b6de00/setup.cfg#L71)
and because the warning that was catching was [removed in scipy
1.4.0](https://github.com/scipy/scipy/issues/10395) (thanks @jni!).

The one in the `Layer.thumbnail` setter is a little harder to justify,
but here's my reasoning.
- I protected one common case of warnings, which is non-finite values.
- Even if we get a warning, it will be shown once by default.
- Since we depend on scikit-image now, I considered just using its
`img_to_ubyte` instead of our `convert_to_uint8`, but I think it warns
in more cases.

Obviously, we have other usage of `catch_warnings` that could cause
similar problems and this PR doesn't touch those. In general, we should
be cautious about using `catch_warnings` in code paths that are hit
frequently.

## Type of change
- [x] Bug-fix (non-breaking change which fixes an issue)

# References
Closes #5588.

# How has this been tested?
- [x] all existing tests pass with my change

I manually tested with `examples/add_multiscale_image`.