FIX: fix translation checking script.
The script was not handling string with more than one letter prefix,
in more details, it was trimming only the first letter of the prefix,
and thus when evaluating for example a rf"" f-string, the f would not be
removed, and the eval would look try to interpolate the ftrings in the
context of the function that then would fail.

Now we keep the prefix, but remove "f" from it.

As a second side effect, this also mean that we now distinguish
bytestrings (`b""`), anc can just skip over those as they should never
be translated.

Another addition in the interactive translation updating script, is that
it now print at the beginning an estimate of the number of places that
are missing translation or need to be added to the ignore file.

The translation tests will still fail, but this is because there are
missing translation/entries in the ignore file.

I did not update he ignore file yet to keep things reviewable.