Fix waiting for multiple signals in threading progress bar test (#5637)
# Description
This fixes some syntax used to wait for multiple Qt signals in a test
related to Qt thread workers. I don't think using `and` to apply two
context managers does what we want (i.e. waiting for both signals to be
emitted), but happy to be corrected. I think we could use `,`, but since
`qtbot` provides `waitSignals` I went with that.

## Type of change
- [x] Bug-fix (non-breaking change which fixes an issue)

# References

This [test failed for me in an related
PR](https://github.com/napari/napari/actions/runs/4420615947/jobs/7750535066?pr=5631#step:8:408).
I'm not sure that this change will prevent that test from failing in the
future, but it should at least help.

# How has this been tested?
- [x] the changed test continues to pass locally