Do not use cgi-traceback on 3.11+
cgitb that is used as a fallback to convert traceback to html has been
deprecated as part of Python 3.11 and is marked for removal in python
3.13.

As we turn warning into errors in Python 3.11+ this will make the test
fail.

This add a conditional to simply wrap the traceback into pre tags when
IPython is not available and we are on Python 3.11+.

I would be tempted to already completely remove cgi-tb code for us to
not have a difference in behavior before/after 3.11.