Bugfix: Convert Viewer Delete button to QtViewerPushButton with action and shortcut (#5636)
The current trash-icon delete button in the viewer when clicked deletes
selected layers, but is also supposed to accept drag-n-drop events,
making it work like the Trash on say macOS (and other OS with a similar
desktop model).
However, this behavior was broken in
https://github.com/napari/napari/pull/2441

This (broken) behavior makes no sense, because napari *doesn't use* a
desktop model, so the Trash icon is not a folder, like the Trash bin in
a OS, a temporary storage for unwanted items. You can't open the napari
trash button to restore the layers you drag there.

The Trash icon in napari should be simply a button—and is stylized as
one—that responds to a single-click to delete selected layers.

This PR removes the broken behavior—eliminating the source of error. In
this way it simplifies the code, to treat the delete button the same as
other viewer buttons (QtViewerPushButton). as a result, the delete
button now also has a registered action and keybinding that is editable
in the Preferences, rather than hard-coded.

<!-- Please delete options that are not relevant. -->
- [x] Bug-fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected): it's been broken for 2 years
without being reported...
- [x] This change requires a documentation update: docs PR:
https://github.com/napari/docs/pull/123

closes https://github.com/napari/napari/issues/5629
See also discussion on zulip:
https://napari.zulipchat.com/#narrow/stream/212875-general/topic/Drag.20layer.20to.20trash

<!-- Please describe the tests that you ran to verify your changes. -->
- [ ] example: the test suite for my feature covers cases x, y, and z
- [ ] example: all tests pass with my change
- [ ] example: I check if my changes works with both PySide and PyQt
backends
      as there are small differences between the two Qt bindings.

- [x] My PR is the minimum possible work for the desired functionality
- [ ] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation: docs PR
https://github.com/napari/docs/pull/123
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] If I included new strings, I have used `trans.` to make them
localizable.
For more information see our [translations
guide](https://napari.org/developers/translations.html).