Bugfix: documentation update for napari PR #5636 (#123)
# Description

Currently, the Viewer tutorial describes the delete button (trash) as
being able to accept drag-n-drop events, making it work like the Trash
on say macOS (and other OS with a similar desktop model). However, this
behavior was broken in https://github.com/napari/napari/pull/2441

It's not intuitive, so the broken behavior is being removed in napari PR
[#5636](https://github.com/napari/napari/pull/5636)
This companion PR updates the Viewer tutorial for this change.

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] Fixes or improves existing content
- [ ] Adds new content page(s)
- [ ] Fixes or improves workflow, documentation build or deployment

# References
Companion to https://github.com/napari/napari/pull/5636

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added [alt text](https://webaim.org/techniques/alttext/) to
new images included in this PR