MAINT: Don't test windows-bundle on every PR.
Windows bundle is one of the slowest (20min),
Let's not test that on every commit.

It would be better to only test bundle if other jobs are succoring,
but this seem to be quite complicated to do.

Therefore we try to do that only by testing whether we are on a PR.