Fix features issues with init param and property setter (#5646)
# Description
This introduces a `feature_defaults` init parameter and property setter
for the three layer types that support that (points, shapes, and
vectors). It also fixes some related issues that occur when adding
feature-derived colors when modifying the feature defaults in-place.

Note that the `feature_defaults` is deliberately a little different to
the `current_properties` setter, as the latter also has the side-effect
of setting the feature values of the selected elements to the new
current property values. Also, using the `feature_defaults` setter with
keys/columns that are extra/missing compared to `features` will result
in an error.

## Type of change
- [x] Bug-fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)

# References
Closes #5632
Closes #5634

# How has this been tested?
- [x] added new tests to cover bugs and setter
- [x] all tests pass with my change