Update PR template to front-load issue references (#5659)
# Description

Currently, when PRs have long and detailed descriptions (a good
thing!!!), it's
hard to find the reference to the issue(s) that they close. This updates
the
template so that those issues are linked right at the top, before the
description.

I think a lot of the stuff at the bottom of the PR template is cruft (it
was
created a long time ago), but we can discuss that in separate
issues/PRs.