Thick Slices (Dims  as nD-box instead of nD-point) (#5522)
# Description
Finally reviving this old code to upgrade our `Dims` model from a single
point to an n-dimensional bounding box.

The basic change is that instead of defining the current `Dims`
"position" by just an nD point (n-tuple), it is instead defined by a
point *and* a left and right margin for each dimension. These margins
describe the (asymmetric) thickness of the visible bounding box starting
from the `point` in world coordinates.

In this PR, this information is not yet used by layers nor exposed via
gui. However, in later PRs we can use this to project data with higher
dimensionality than `ndisplay` contained in the bounding box onto the
visible space.

See #5697 for a possible implementation and videos of the thick slicing.

---

With this PR, the main "primary" fields for `Dims` look like this:

Same as before:
- `ndim`
- `ndisplay`
- `order`
- `axis_labels`

Changed:
- `point`: no longer a property, but the source of truth for the dims
position. Expressed in wolrd coordinates.
- `current_step`: no longer a field, but a computed property
- `range`: now a named tuple for ease of use. One significant change
here is that `range.stop` used to be an excluded bound (probably to
mirror the builtin `range` function). However, it only made sense when
the source of truth was integers (`current_step`) instead of continuous
space (`point`). This change can be seen in some of the tests.

Added:
- `margin_left`: thickness on the left/lower side of the `point` for
each dimension, in *world space*
- `margin_right`: same as above, but to the right/higher side
- `thickness`: a shortcut to `margin_left` and `margin_right` that
assumes equally thick margins of `thickness / 2` width.

There are also a few changes to properties and methods to streamline the
code and usage.

Changed properties:
- `nsteps`: now settable (will change `step` to match the requested
number of steps)
- `displayed`: can now be set (will affect `self.order`)

---

Other relevant notes:
- now everything is in world space, and anything which is not gets
computed as a property (differently from main, where `current_step` is
in "slider space" and is the primary source of the dims state).
- there were many previous iterations and a lot of discussion around the
implementation (particularly around using evented containers for a more
ergonomic interface). Please refer to the Reference section below for a
complete history, but **TL;DR: nested evented containers with non-buggy
validation are impossible or unreasonably slow with our current
`EventedModel`**.

## Type of change
<!-- Please delete options that are not relevant. -->
- [ ] Bug-fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] This change requires a documentation update

# References
- original implementation (which ultimately is very close to this :P):
#4012
- example of how layers could be hooked up to it:
https://github.com/napari/napari/pull/4012#issuecomment-1049850169 and
https://github.com/napari/napari/pull/4012#issuecomment-1057098537
- broken version with evented containers: #4334 
- some extra stuff that didn't work out to make working with evented
objects and property better:
	- #4417 
	- #4457
- ... and the "Thick slices" PR that came out of these: #4448, #4522
- attempts at getting nested evented models to work:
	- #4474 
	- #4597 
	- #4609 
	- #4804

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>