Zooming with the mouse wheel in any mode (#5701)
# Fixes/Closes

Closes #4837

# Description
## Why the current behavior is not very convenient
The current implementation of napari layers and camera disables zooming
and panning with the mouse wheel in any mode other than `PAN_ZOOM`. This
behavior is very inconvenient for annotating given the fact that the
mouse wheel does nothing in other modes. Holding space to temporarily
switch to the `PAN_ZOOM` mode does not always help because, for example,
if you are in the process of creating a polygon it forces to prematurely
finish it.

## Source of the problem
Napari does not have its own implementation of zooming and panning and
utilizes the functionality of VisPy cameras to achieve that. So, the
above-mentioned behavior actually comes from the VisPy implementation of
SceneCanvas and its Cameras by changing the `interactive` attribute.
Setting it to `False` disables the mouse events completely without
distinguishing between pan and zoom in two different places of VisPy:
1. `vispy.scene.cameras.panzoom.PanZoomCamera` (or ArcballCamera for 3D)
blocks processing the mouse events if it is set as `interactive =
False`:

https://github.com/vispy/vispy/blob/5449e3402b35da4d6a8215d8d72e19dcf05917d7/vispy/scene/cameras/panzoom.py#L200-L201
2. `vispy.scene.canvas.SceneCanvas` in the method `_process_mouse_event`
(through calling `visual_at` and `_visual_bounds_at`) checks whether the
node is interactive or not, and if it is not, it prevents mouse events
from being passed from the canvas to the viewbox containing the VisPy
camera. In practice, when `interactive = False`, we never even reach the
condition above.

https://github.com/vispy/vispy/blob/5449e3402b35da4d6a8215d8d72e19dcf05917d7/vispy/scene/canvas.py#L425-L427

https://github.com/vispy/vispy/blob/5449e3402b35da4d6a8215d8d72e19dcf05917d7/vispy/scene/canvas.py#L345-L358

This all makes it tricky to decouple zooming and panning while being
able to disable the default mouse events when one wants to implement
custom mouse behavior in other modes. That's why napari has this strange
behavior when the mouse wheel is free and no other modes can use it for
zooming.

## Minimally invasive solution

First of all, in order to selectively deactivate panning and zooming, we
should refuse to set `interactive=False` for the VisPy objects, as doing
so blocks passing mouse events inside the SceneCanvas and there is
nothing we can do about it. Second, the way how PanZoomCamera and
ArcballCamera handle mouse events should be changed.

In this PR:
1. I deprecated the `interactive` attribute from all napari objects and
replaced it with two new attributes `mouse_pan` and `mouse_zoom`, that
determine if panning or zooming for a layer or camera is activated. (by
the way, this naming `interactive` is very confusing as it is too
abstract and does not refer to the mouse in any way, it took me a good
few hours with a debugger to find out what really affects mouse events).
The deprecated `interactive` attribute will have the same functionality
as `mouse_pan`.
2. The behavior of the PanZoomCamera/ArcballCamera mouse callbacks is
changed by introducing the class decorator `add_mouse_pan_zoom_toggles`
that takes any subclass of BaseCamera and adds `mouse_pan` and
`mouse_zoom` attributes to it, which allows to selectively block mouse
move/click and wheel events even when `interactive = True`.
3. All the old use cases of the `interactive` attribute are replaced
with the new `mouse_pan` attribute. It can be done safely because, for
all I know, they only care about disabling the mouse dragging/panning,
so that to free up mouse move and click events.
4. `PAN_ZOOM` mode has `mouse_pan = True` and `mouse_zoom = True`,
switching to other modes only disables `mouse_pan`, preserving the
ability to zoom with the mouse wheel.


# References
<!-- What resources, documentation, and guides were used in the creation
of this PR? -->

## Type of change
- [x] New feature (non-breaking change which adds functionality)

# How has this been tested?
I tested it manually using different annotation layers in the 2D in 3D
camera modes.

## Final checklist:
- [X] My PR is the minimum possible work for the desired functionality
- [X] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] If I included new strings, I have used `trans.` to make them
localizable.
For more information see our [translations
guide](https://napari.org/developers/translations.html).

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: alisterburt <alisterburt@gmail.com>