Unify event behaviour for points and its qt controls (#5722)
# Fixes/Closes
Closes #5718.

# Description
Point controls were not firing events as intended. Most of the event
blockers were either too aggressive (block everything, rather than just
the relevant callback) or inverted.

Additionally, there was a lot of asymmetry and confusion between how
various elements worked. I clarified the gui labels to make it clearer
when the `current_*` property is being affected. I also added a few
events to match expectations with other parts of the `Points` layer.

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] Bug-fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update