Feature: Background color swapping for the Label layer (extends #5672) (#5744)
# Description

#5672 introduced a new feature allowing switching to the background
color. To make it more useful, I propose extending the behavior of that
feature by implementing the swapping behavior, not just setting to the
background color:

- If the current label is not the background label, it will be set to
the background label.
- If the current label is the background label, it will be set to the
previous non-background label.

Actually, this is a standard functionality of almost any image editor
such as GIMP or Photoshop. It is really handy when you are editing masks
as it allows quick switching between adding and removing.

I also changed the hotkey from `B` to `X` to match the default hotkeys
of GIMP and Photoshop that do exactly the same thing, which should make
it more convenient for people who are used to working with similar
software.

## Type of change
- [X] Breaking change (changes the behavior of #5672)

# How has this been tested?
- [X] added test for the background swapping
- [X] all tests pass with my change

## Final checklist:
- [X] My PR is the minimum possible work for the desired functionality
- [X] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] If I included new strings, I have used `trans.` to make them
localizable.
For more information see our [translations
guide](https://napari.org/developers/translations.html).