Disable buffer swapping (#5741)
# Fixes/Closes
Closes #5734.

# Description
This change drastically improves performance on some hardware by
sacrificing some quality (tearing is more likely). However, several
people tested it and didn't notice any appreciable degradation of the
rendering, so it seems like a worthwile exchange. See #5734 for more
details.

We might be able to fix this on the vispy side, but for now this works
well on the napari side.

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] Bug-fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

cc @ksofiyuk

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>