Refactor: Qt viewer and VispyCanvas  (#5432)
# Description
This is a PR in preparation of multi canvas view as discussed in #5348.
It is a refactor of mainly the QtViewer and the Vispy Canvas. A lot of
the Vispy functionality was within the Qt viewer, while not granting a
way for example for scene parents to be adjusted on the Vispy side later
on. For implementation of the multicanvas viewer it is required to have
more control on the Vispy side.
Furthermore, ensuring that Qt and Vispy work as much as possible in
isolation also ensures that the codebase of the 2 is not too entangled
and thus allows for switching to other back or frontends more easily.

With this PR, all the VispyCanvas calls have been moved to VispyCanvas,
except for key events and mouse events. Rerouting key events is already
discussed in another PR (#5296). For mouse events a PR will be opened at
a later stage to allow napari to handle mouse events happening on
different canvases and keeping track of them. This would require
significant changes and thus would be better handled in a separate PR.
The PR does not affect the user API. Also, backward compatibility has
been added to take into account plugin developers who have based their
plugins on the current codebase.

Regarding the tests, these all have been adjusted to reflect the code
changes. However, it could be worth discussing whether certain tests
would be better to move to _vispy/_tests instead of _qt/_tests.

## Type of change
- [x] Refactor

# References
This PR is a prerequisite for  #5348

# How has this been tested?
Locally using pytest. All tests passed. To the best of my knowledge the
_process_mouse_event fix discussed in PR #1798 is not covered by tests.
I tried recreating the issue to see whether the fix was still required,
but could not recreate. To be certain the code fix has been maintained
by
[dda628d](https://github.com/napari/napari/pull/5432/commits/dda628d7a8d7c68e12daac3f53a777c810023418),
but it could be worth checking by others whether this fix is still
required.

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests 
- [x] If I included new strings, I have used `trans.` to make them
localizable.
For more information see our [translations
guide](https://napari.org/developers/translations.html).

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>