Prevent canvas micro-panning on point add (#5742)
Currently, when adding points interactively the vispy camera is still
getting moved by mouse events.

This is not a problem with a mouse - since normally you don't move the
mouse while clicking - but it's pretty frustrating with a tablet, since
the cursor moves slightly every time you press down or release the
stylus. This results on the canvas moving around a tiny bit every time
you click.

This was a know issue, since there is a mouse-movement threshold in
place beyond which a point doesn't get added; however this does not
prevent the canvas from moving even within that threshold.

This PR prevents the vispy camera from receiving the event if this
threshold is not yet surpassed.

<!-- Please delete options that are not relevant. -->
- [x] Bug-ix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update