Throttle mouse move event to fix high-polling-rate input (#5710)
# Fixes/Closes

This is a solution for #5709. It seems that the polling rate of some
input devices (like the tablet+stylus I'm testing) is too high for our
code to handle. This issue is especially bad on machines with worse
CPUs; the one I have is actually quite decent, but on another machine
this issue was crippling even with the smallest shapes.

# Description

Solves the issue by throttling all `mouse_move` events to 5ms, which
ensures that we don't process every single event we receive, no matter
how fast they come.

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] Bug-fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update