Add cancellation functionality to progress (#5728)
# Description

This PR provides the ability for users to cancel tasks utilizing the
`progress` mechanism. Therefore, if you write something like

```python
for i in progress(range(10)):
        do_something_time_consuming()
```
and the user decides to cancel on the fifth iteration of the `for` loop,
the remaining 5 iterations will **not** be run. The user can assert
cancellation with the field `progress.is_canceled()`, and can provide
cancellation error-handling by passing a `Callable` to the `progress`
object as the `cancel_callback` parameter - this can enable context
destruction in the case of cancellation.

# References
<!-- What resources, documentation, and guides were used in the creation
of this PR? -->

## Type of change
<!-- Please delete options that are not relevant. -->
- [X] New feature (non-breaking change which adds functionality)
- [X] This change requires a documentation update

# How has this been tested?
<!-- Please describe the tests that you ran to verify your changes. -->
- [X] test suite for my feature covers cases x, y, and z
- [X] all tests pass with my change

## Final checklist:
- [X] My PR is the minimum possible work for the desired functionality
- [X] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] If I included new strings, I have used `trans.` to make them
localizable.
For more information see our [translations
guide](https://napari.org/developers/translations.html).

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Draga Doncila Pop <17995243+DragaDoncila@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>