Changing brush size with mouse movement while holding specified modifiers (#5753)
# Description

Currently, adjusting the brush size in napari requires using keyboard
shortcuts, which can be inconvenient and disrupt the workflow as, for
example, you need to remove your right hand from the mouse to reach the
`[` and `]` keys, let alone it takes a lot of time of holding `]` to
increase the brush size from a very small size to a big one.

This PR provides a more intuitive way to adjust the brush size, making
it easier and faster to choose the desired brush size. When a user moves
the mouse while holding the 'Alt' key (it can be changed in the
preferences), the callback adjusts the brush size based on the direction
and distance of the horizontal mouse movement. The amount of adjustment
is proportional to the distance the mouse is moved.

It mimics the way the brush size can be controlled in
[CVAT](https://www.cvat.ai/), which I found to be a very handy and
intuitive way of adjusting the brush size.


https://user-images.githubusercontent.com/5459767/234982700-1ef05508-6cb0-44e5-a3df-ca051b0b7bdc.mp4

~~Unfortunately, you can see that the mouse is still slightly shaking,
but I can't see a good way how to address this as it happens because of
the high latency of processing callbacks in Python - it may take up to
ten milliseconds from the moment the system registered the mouse
movement to the moment when it is processed by the callback in napari.
It doesn't look perfect, however it doesn't disturb too much as well.
When the overlay for the brush visualization is implemented, it will be
possible to address this issue properly.~~ Fixed by using the brush
circle overlay
[here](https://github.com/napari/napari/pull/5753#issuecomment-1526389602).


A new setting to Preferences->Application is added that controls what
modifiers activate the callback (`Alt` (default), `Control`,
`Control+Alt`, `Control+Shift`) or just disable it:
 

![image](https://user-images.githubusercontent.com/5459767/235266826-0fab4a43-6578-46f1-88f9-92d760876dfa.png)

## Type of change
<!-- Please delete options that are not relevant. -->
- [X] New feature (non-breaking change which adds functionality)

# How has this been tested?
<!-- Please describe the tests that you ran to verify your changes. -->
- [X] Tested it manually 

## Final checklist:
- [X] My PR is the minimum possible work for the desired functionality
- [X] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [X] If I included new strings, I have used `trans.` to make them
localizable.
For more information see our [translations
guide](https://napari.org/developers/translations.html).

---------

Co-authored-by: alisterburt <alisterburt@gmail.com>